==> 0_data_collection.ipynb <==
import pandas as pd
import twint
from tqdm import tqdm
from textblob import TextBlob

We run following cell just to test the code for a small output (my own username,  10 tweets)

# c = twint.Config()
# c.Username = "luj_fel"
# c.Custom["tweet"] = ["id"]
# c.Custom["user"] = ["bio"]
# c.Limit = 10
# c.Store_csv = True
# c.Output = "none"
# twint.run.Search(c)

Define a function to supress output

import sys, os

# Disable
def blockPrint():
    sys.stdout = open(os.devnull, 'w')

# Restore
def enablePrint():
    sys.stdout = sys.__stdout__

print('This will print')

blockPrint()
print("This won't")

enablePrint()
print("This will too")

Read in the saved usernames of PIF csv

x = 'usernames.csv'
x = pd.read_csv(x)
# x = list(x.username)

x.shape

x.drop_duplicates(inplace=True)

check the number of followers usernames

x.shape

x = list(x.username)

x[:10]

Fetch the last 20 tweets for each username in the list

Notes: 


- The limit in the code was set to 10, however the minimum for twint is 20

- The code below was run on an AWS instance

import twint
import pandas as pd

x = 'usernames.csv'
x = pd.read_csv(x)
x.drop_duplicates(inplace=True)
x = list(x.username)

for a in x:
    v = twint.Config()
    v.Username = a
    v.Limit = 10
    v.Store_csv = True
    v.Output = "tweets.csv"
    v.Custom["tweet"] = ["username", "tweet"]
    twint.run.Search(v)
