==> 1_en_users_extract.ipynb <==
import pandas as pd
from tqdm.autonotebook import tqdm
tqdm.pandas()

from langdetect import detect

detect("ليش نحنا نحب الاكل")

def lang(x):
    try:
        return detect(x)
    except:
        return a

Read in the fetched 20 tweets dataframe of all PIF followers 

ef = pd.read_csv('tweets_ver3.csv')
ef.info()

Detect the Language for each tweet (observation)

ef['lang'] = ef.tweet.progress_apply(lambda x : lang(x) == 'en')

Group by username, then sum the language values

ef2 = ef.groupby(['username']).sum()

ef2.info()

Filter by value of language sume, if 75% of all recent tweets are english, we assume the user is english

ef3 = ef2[ef2['lang'] >14]

ef3.info()

We have 600+ users who recently tweets mostly in english down from 40K total users, we assume the chosen data sample is respresentative of all followers. We only choose english to avoid Arabic NLP limitation for now

ef3.head()

ef3.reset_index(level=0, inplace=True)

Save teh list of EN users to a CSV File

ef3[['username']].to_csv('enusers_finale.csv', index=False, header=True)

pd.read_csv('enusers_finale.csv')

Now that we have the list of english users, we fetch ALL of their tweets (max 1500 per user)

import twint
import pandas as pd

x = 'enusers_finale.csv'
x = pd.read_csv(x)
x = list(x.username)

for a in x:
    v = twint.Config()
    v.Username = a
    v.Limit = 1500
#     v.Custom["tweet"] = ['id', 'date', 'time', 'timezone', 'username', 'tweet', 'mentions', 'location', 'hashtags', 'retweet']
    v.Store_csv = True
    v.Output = "tweets_finale_full.csv"
    twint.run.Search(v)
