==> 2_data_filtering.ipynb <==
import pandas as pd
pd.set_option('display.max_columns', None)
import matplotlib.pyplot as plt
import seaborn as sns
import numpy as np

from tqdm.autonotebook import tqdm
tqdm.pandas()
from langdetect import detect

Read the data

df = pd.read_csv('tweets_finale_full.csv', low_memory=False)

df.info()

df.head(2)

TO MAKE SURE WE SCRAPED ALL THE USERS!

us = list(df.username.unique())

# li = list(li.username)

# oh = list(set(us).intersection(li))

len(oh)

The dif between the wto lists is 1, so we're mostly good

len([item for item in li if item not in us])

# check id is unique
df.id.nunique()

# set id to be index
df.set_index('id', inplace=True)

# define lang detect function
def lang(x):
    try:
        a = detect(x)
    except:
        a = 0
    return a

Apply language detection function to the entire dataframe

df['lang'] = df.tweet.progress_apply(lambda x : 1 if lang(x) == 'en' else 0)

df.to_csv('tweets_finale_full_lang.csv')

df.username.nunique()

df_easy = pd.read_csv('tweets_finale_full_lang.csv')

df_easy.head()

df_easy.info()

df_easy = df.drop(columns = ['conversation_id', 'created_at', 'user_id', 'name', 'urls', 'photos',
                             'location', 'link', 'retweet', 'place'] , axis=1)

df_easy.head()

df_easy.info()

df_easy.to_csv('tweets_finale_full_lang_dropped.csv')

af = pd.read_csv('tweets_finale_full_lang_dropped.csv')

af.drop_duplicates(inplace=True)

af.info()

af.head()

af.username.nunique()

Make sure we only have tweets of en usernames

bf = af[af.username.isin(uslist)]
bf.head()

bf.to_csv('tweets_finale_finale2_lang.csv')

bf.info()

af.head()

Checking language for tweets before starting to clean

import re

Clean the tweets a bit before cleaning to make sure URLS only tweets are not detected as EN

re.sub('pic.twitter.com\S+', '', str('pic.twitter.com/oqW8eeHfwg'))

af.tweet.map(lambda x : re.sub('pic.twitter.com\S+', '', str(x))) == ' '

bf.shape

bf = bf[bf.tweet.str.replace('\S*pic.twitter.com\S*', '', regex=True) != ' ']

bf.shape

bf.info()

Define username list of users who we have at least 75 english tweets for

ef = bf[['username', 'lang']]

gf = ef.groupby(['username']).sum()

gf2 = gf[gf['lang'] >75]

gf2.info()

# gf2.reset_index(level=0, inplace=True)
# gf2[['username']].to_csv('finale_finale_enusers.csv', index=False, header=True)
uslist = pd.read_csv('finale_finale_enusers.csv')
uslist = list(uslist.username)
len(uslist)

filter original data frame by english users from filtered list

af = af[af.username.isin(uslist)]
af.head()

af.info()

Filter to only include english tweets

af = af[af.lang == 1]

af.info()

Save final dataframe

af.to_csv('finale_go_clean.csv', index=False, header=True)
