==> 5_data_basic_modeling.ipynb <==
import pandas as pd
import numpy as np
import re
import matplotlib as plt
from wordcloud import WordCloud
from tqdm.autonotebook import tqdm
tqdm.pandas()

# NLTK
from nltk.tokenize import RegexpTokenizer
from nltk.stem import PorterStemmer
from nltk.corpus import stopwords

# Visualization
import matplotlib.pyplot as plt
import matplotlib.patches as mpatches
import matplotlib
%matplotlib inline
import seaborn as sns

gf = pd.read_csv('finale_go_clean.csv')

ff['sentiment'] = gf['tweet'].progress_apply(lambda x: TextBlob(x).sentiment.polarity)

ff.head()

ff['Timestamp'] = pd.to_datetime(ff['date'] + ' ' + ff['time'])

ff['DayOfWeek'] = ff['Timestamp'].progress_apply(lambda x: x.dayofweek)

ff['Month'] = ff['Timestamp'].progress_apply(lambda x: x.month)

ff['Hour'] = ff['Timestamp'].progress_apply(lambda x: x.hour)

ff.head()

mf = ff[['id', 'username', '0', '1', '2', '3', '4', '5', '6', '7', '8', '9', 'DayOfWeek', 'Month', 'Hour', 'sentiment']]

# Preliminary Prediction Models

X = mf[['username', '0', '1', '2', '3', '4', '5', '6', '7', '8', '9', 'DayOfWeek', 'Month', 'Hour']]
X = pd.get_dummies(X)

X = pd.get_dummies(X)

y = mf[['sentiment']]

from sklearn.preprocessing import PolynomialFeatures, StandardScaler
from sklearn.linear_model import LinearRegression, Ridge, Lasso, RidgeCV, LassoCV
from sklearn.model_selection import cross_val_score, cross_val_predict,train_test_split
from sklearn.metrics import r2_score, mean_squared_error
import matplotlib
import matplotlib.pyplot as plt

X_train, X_test, y_train, y_test = train_test_split(X,y,test_size=0.2)

scaler = StandardScaler()
X_train = pd.DataFrame(scaler.fit_transform(X_train), columns=X.columns)
X_test = pd.DataFrame(scaler.transform(X_test), columns=X.columns)

model_lr = LinearRegression()
scores = cross_val_score(model_lr,X_train,y_train,cv=5)
print('scores: ',scores)
print('socres mean: ',scores.mean())

from sklearn.tree import DecisionTreeRegressor

dt = DecisionTreeRegressor()
dt.fit(X_train, y_train)

print('training           r2:',dt.score(X_train,y_train))
print('test               r2:',dt.score(X_test,y_test)) 

from sklearn.ensemble import GradientBoostingRegressor,RandomForestRegressor,BaggingRegressor,AdaBoostRegressor

a= RandomForestRegressor(max_features=12)
a.fit(X_train,y_train)

print('training           r2:',a.score(X_train,y_train))
print('test               r2:',a.score(X_test,y_test))

sc.mean(ff.sent_class)

CLASSIFICATION

import matplotlib.pyplot as plt
from sklearn.linear_model import LogisticRegression

def osa(x):
    if x > 0.2:
        o = 2
    elif 0.2 > x > -0.2:
        o = 1
    else:
        o = 0
    return o 

ff['sent_class'] = ff['sentiment'].progress_apply(lambda x: osa(x))

yc = ff[['sent_class']]

X_train_c, X_test_c, y_train_c, y_test_c = train_test_split(X,yc,test_size=0.2)

scaler = StandardScaler()
X_train_c = pd.DataFrame(scaler.fit_transform(X_train_c), columns=X.columns)
X_test_c = pd.DataFrame(scaler.transform(X_test_c), columns=X.columns)

import matplotlib.pyplot as plt
from sklearn.linear_model import LogisticRegression

logreg = LogisticRegression(solver='lbfgs', multi_class='multinomial')
logreg.fit(X_train_c, y_train_c)
logreg.score(X_train_c, y_train_c), logreg.score(X_test_c, y_test_c)

import scipy as sc

from sklearn.tree import DecisionTreeClassifier

tree = DecisionTreeClassifier(max_depth=20, min_samples_split=4)
tree.fit(X_train_c, y_train_c)
tree.score(X_train_c, y_train_c), tree.score(X_test_c, y_test_c)

yc.sent_class.value_counts()

241778/(len(yc))

# from imblearn.over_sampling import SMOTE
# Xres_train,Xres_test , yres_train, yres_test = train_test_split(X, y, test_size = 0.25)


# smote = SMOTE()
# Xres_train, yres_train = smote.fit_resample(Xres_train, yres_train)

# from sklearn.linear_model import LogisticRegressionCV

# model_res = LogisticRegression()
# model_res.fit(Xres_train, yres_train) 
# model_res.score(Xres_train, yres_train), model_res.score(Xres_test, yres_test)

# pd.Series(yres_train).value_counts()

df = ff[['id', 'username', '0', '1', '2', '3', '4', '5', '6', '7', '8', '9', 'DayOfWeek', 'Month', 'Hour', 'sentiment', 'sent_class']]

df.head()

# df.to_csv('FINAL_DATAFRAME_CLASS.csv')

df_drop = df[df.sent_class != 1]

df_drop.username.nunique()

df_drop.info()

X = df_drop[['username', '0', '1', '2', '3', '4', '5', '6', '7', '8', '9', 'DayOfWeek', 'Month', 'Hour']]
X = pd.get_dummies(X)

scaler = StandardScaler()
X = pd.DataFrame(scaler.fit_transform(X), columns=X.columns)

y = df_drop[['sentiment']]

yc = df_drop[['sent_class']]

model_lr = LinearRegression()
scores = cross_val_score(model_lr,X,y,cv=5)
print('scores: ',scores)
print('socres mean: ',scores.mean())

logreg = LogisticRegression(solver='lbfgs', multi_class='multinomial')
scores = cross_val_score(logreg,X,yc,cv=5)

scores
