==> 6_data_heavy_modeling.ipynb <==
import pandas as pd
import seaborn as sns
import numpy as np

# Modelling

df = pd.read_csv('FINAL_DATAFRAME.csv')

df.head()

af = df[['id', 'username', '0', '1', '2', '3', '4', '5', '6', '7', '8', '9', 'DayOfWeek', 'Month', 'Hour', 'sentiment']]

af.head()

import matplotlib.pyplot as plt

kf =  pd.read_csv('FINAL_DATAFRAME_CLASS.csv')

kf.drop(['Unnamed: 0'], axis=1, inplace=True)

kf.head()

from sklearn.preprocessing import PolynomialFeatures, StandardScaler

X = kf[['username', '0', '1', '2', '3', '4', '5', '6', '7', '8', '9', 'DayOfWeek', 'Month', 'Hour']]
X = pd.get_dummies(X)

ss = StandardScaler()
X = pd.DataFrame(ss.fit_transform(X), columns=X.columns)
# I recognize here that I should've scaled after train test split.

y = kf[['sentiment']]

yc = kf[['sent_class']]

from sklearn.model_selection import train_test_split
X_train, X_test, yc_train, yc_test = train_test_split(X,yc,test_size=0.2)

from sklearn.neural_network import MLPClassifier

clf = MLPClassifier(solver='lbfgs', alpha=1e-5, hidden_layer_sizes=(5, 5, 5))

clf.fit(X_train, yc_train)                         
clf.score(X_train, yc_train), clf.score(X_test, yc_test)    

def one_hot_encoder(array):
    unique_values, indices = np.unique(array, return_inverse=True)
    print(unique_values)
    print(indices)
    one_hot_encoded_data = np_utils.to_categorical(indices, len(unique_values))
    return one_hot_encoded_data

from keras.models import Sequential
from keras.layers import Dense, Conv2D
from keras.utils import np_utils

X.shape

yc_train_encoded = one_hot_encoder(yc_train)

yc_test_encoded = one_hot_encoder(yc_test)

model = Sequential()
model.add(Dense(128, input_shape=(442,), name='Input_Layer', activation='relu'))
model.add(Dense(256, activation='relu'))
# model.add(Dense(64, activation='relu'))
model.add(Dense(3, name='Output_Layer', activation='softmax'))
model.summary()

model.compile(optimizer='Adam', loss='categorical_crossentropy', metrics=['accuracy'])

model.fit(X_train, yc_train_encoded, batch_size=None, epochs=30, steps_per_epoch=5)

model.evaluate(X_test, yc_test_encoded)

X_test.shape

model.save('lujain_model.h5')
