==> 8_case_predict.ipynb <==
import pandas as pd
import seaborn as sns
import numpy as np

# Using Trained Keras Model

model.save('lujain_model.h5')

# Case Prediction

f = pd.read_csv('Predicting_Set.csv')

f.head()

f.drop(['Unnamed: 0'], axis=1, inplace=True)

gf = pd.get_dummies(f)

pred = model.predict(gf)

pred.shape

pred_df = pd.DataFrame(pred)

pred_df.columns = ['class_0', 'class_1', 'class_2']

pred_df.head()

full_df = f.join(pred_df, how='inner')

full_df.head()

full_df.to_csv('predicted_set.csv')
